dataset/multiple_commits/BskyBot/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
atproto==0.0.56
beautifulsoup4==4.12.3
blueskysocial==1.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==43.0.3
dnspython==2.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
libipld==3.0.0
pillow==11.0.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
pyUnicodeSteganography==0.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
steganography==0.1.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
